+++ interpreter/namerd/src/main/scala/io/buoyant/namerd/iface/NamerdInterpreterInitializer.scala
@@ -2,13 +2,16 @@ package io.buoyant.namerd.iface
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.twitter.conversions.time._
+import com.twitter.finagle.Stack.Transformer
 import com.twitter.finagle._
 import com.twitter.finagle.naming.NameInterpreter
 import com.twitter.finagle.param.HighResTimer
 import com.twitter.finagle.service._
+import com.twitter.logging.Logger
 import com.twitter.util._
 import io.buoyant.namer.{InterpreterConfig, InterpreterInitializer}
 import io.buoyant.namerd.iface.{thriftscala => thrift}
+import com.twitter.finagle.buoyant.TlsClientPrep
 
 /**
  * The namerd interpreter offloads the responsibilities of name resolution to
@@ -31,11 +34,16 @@ case class Retry(
   }
 }
 
+case class ClientTlsConfig(commonName: String, caCert: Option[String])
+
 case class NamerdInterpreterConfig(
   dst: Option[Path],
   namespace: Option[String],
-  retry: Option[Retry]
+  retry: Option[Retry],
+  tls: Option[ClientTlsConfig]
 ) extends InterpreterConfig {
+
+  private[this] val log = Logger.get("namerd")
 
   @JsonIgnore
   val defaultRetry = Retry(5, 10.minutes.inSeconds)
@@ -66,7 +74,11 @@ case class NamerdInterpreterConfig(
           def make(_stats: param.Stats, next: ServiceFactory[Req, Rsp]) = {
             val param.Stats(stats) = _stats
             val retry = new RetryFilter[Req, Rsp](
-              RetryPolicy.backoff(backoffs) { case (_, Throw(NonFatal(ex))) => true },
+              RetryPolicy.backoff(backoffs) {
+                case (_, Throw(NonFatal(ex))) =>
+                  log.error(ex, "namerd request failed")
+                  true
+              },
               HighResTimer.Default,
               stats,
               RetryBudget.Infinite
@@ -76,9 +88,20 @@ case class NamerdInterpreterConfig(
         }
     }
 
+    val tlsTransformer = new Transformer {
+      override def apply[Req, Rep](stack: Stack[ServiceFactory[Req, Rep]]) = {
+        tls match {
+          case Some(tlsConfig) =>
+            TlsClientPrep.static[Req, Rep](tlsConfig.commonName, tlsConfig.caCert) +: stack
+          case None => stack
+        }
+      }
+    }
+
     val client = ThriftMux.client
       .withParams(ThriftMux.client.params ++ params)
       .transformed(retryTransformer)
+      .transformed(tlsTransformer)
       .withSessionQualifier.noFailFast
       .withSessionQualifier.noFailureAccrual