+++ linkerd/protocol/http/src/integration/scala/io/buoyant/linkerd/protocol/TlsTerminationTest.scala
@@ -9,7 +9,7 @@ import org.scalatest.FunSuite
 
 class TlsTerminationTest extends FunSuite with Awaits {
 
-  test("tls server + plain backend") {
+  test("netty3 tls server + plain backend") {
     withCerts("linkerd") { certs =>
       val dog = Downstream.const("dogs", "woof")
       try {
@@ -22,6 +22,47 @@ class TlsTerminationTest extends FunSuite with Awaits {
              |    /http/1.1/GET/clifford => /p/dog ;
              |  servers:
              |  - port: 0
+             |    tls:
+             |      certPath: ${certs.serviceCerts("linkerd").cert.getPath}
+             |      keyPath: ${certs.serviceCerts("linkerd").key.getPath}
+             |""".stripMargin
+        val linker = Linker.Initializers(Seq(HttpInitializer)).load(linkerConfig)
+
+        val router = linker.routers.head.initialize()
+        try {
+          val server = router.servers.head.serve()
+          try {
+            val client = upstreamTls(server, "linkerd", certs.caCert)
+            try {
+              val rsp = {
+                val req = Request()
+                req.host = "clifford"
+                await(client(req))
+              }
+              assert(rsp.contentString == "woof")
+
+            } finally await(client.close())
+          } finally await(server.close())
+        } finally await(router.close())
+      } finally await(dog.server.close())
+    }
+  }
+
+  test("netty4 tls server + plain backend") {
+    withCerts("linkerd") { certs =>
+      val dog = Downstream.const("dogs", "woof")
+      try {
+        val linkerConfig =
+          s"""
+             |routers:
+             |- protocol: http
+             |  baseDtab: |
+             |    /p/dog => /$$/inet/127.1/${dog.port} ;
+             |    /http/1.1/GET/clifford => /p/dog ;
+             |  servers:
+             |  - port: 0
+             |    engine:
+             |      kind: netty4
              |    tls:
              |      certPath: ${certs.serviceCerts("linkerd").cert.getPath}
              |      keyPath: ${certs.serviceCerts("linkerd").key.getPath}