+++ project/Base.scala
@@ -17,19 +17,19 @@ import scoverage.ScoverageSbtPlugin
  * Base project configuration.
  */
 class Base extends Build {
-  val headVersion = "0.8.1"
+  val headVersion = "0.8.2"
 
   object Git {
     def git(arg: String, args: String*) = Process("git" +: arg +: args)
     val headRevision = git("rev-parse", "--short", "HEAD").!!.trim
     val version = git("name-rev", "--tags", "--name-only", headRevision).!!.trim match {
-      case tag if tag == headVersion => tag
+      case tag if tag == headVersion || tag == s"${headVersion}^0" => headVersion
       case _ => s"$headVersion-SNAPSHOT"
     }
   }
 
   val developTwitterDeps = settingKey[Boolean]("use SNAPSHOT twitter dependencies")
-  val doDevelopTwitterDeps = developTwitterDeps ?? false
+  val doDevelopTwitterDeps = (developTwitterDeps in Global) ?? false
 
   val baseSettings = Seq(
     organization := "io.buoyant",
@@ -37,16 +37,17 @@ class Base extends Build {
     homepage := Some(url("https://linkerd.io")),
     scalaVersion in GlobalScope := "2.11.7",
     ivyScala := ivyScala.value.map(_.copy(overrideScalaVersion = true)),
-    scalacOptions ++= Seq("-Xfatal-warnings", "-deprecation", "-Ywarn-value-discard"),
+    scalacOptions ++=
+      Seq("-Xfatal-warnings", "-deprecation", "-Ywarn-value-discard", "-feature"),
     // XXX
     //conflictManager := ConflictManager.strict,
     resolvers ++= Seq(
       "twitter-repo" at "https://maven.twttr.com",
-      "local-m2" at s"file:${Path.userHome.absolutePath}/.m2/repository",
+      Resolver.mavenLocal,
       "typesafe" at "https://repo.typesafe.com/typesafe/releases"
     ),
     aggregate in assembly := false,
-    developTwitterDeps := { sys.env.get("TWITTER_DEVELOP") == Some("1") },
+    (developTwitterDeps in Global) := { sys.env.get("TWITTER_DEVELOP") == Some("1") },
 
     // Sonatype publishing
     publishArtifact in Test := false,
@@ -116,6 +117,7 @@ class Base extends Build {
       }),
     assemblyJarName in assembly := s"${name.value}-${version.value}-${configuration.value}-exec",
     assemblyMergeStrategy in assembly := {
+      case "BUILD" => MergeStrategy.discard
       case "com/twitter/common/args/apt/cmdline.arg.info.txt.1" => MergeStrategy.discard
       case "META-INF/io.netty.versions.properties" => MergeStrategy.last
       case path => (assemblyMergeStrategy in assembly).value(path)