+++ project/Deps.scala
@@ -8,15 +8,18 @@ object Deps {
 
   // process lifecycle
   val twitterServer =
-    ("com.twitter" %% "twitter-server" % "1.21.0")
+    ("com.twitter" %% "twitter-server" % "1.24.0")
       .exclude("com.twitter", "finagle-zipkin_2.11")
 
   def twitterUtil(mod: String) =
-    "com.twitter" %% s"util-$mod" % "6.35.0"
+    "com.twitter" %% s"util-$mod" % "6.38.0"
 
   // networking
   def finagle(mod: String) =
-    "com.twitter" %% s"finagle-$mod" % "6.36.0"
+    "com.twitter" %% s"finagle-$mod" % "6.39.0"
+
+  def netty4(mod: String) =
+    "io.netty" % s"netty-$mod" % "4.1.4.Final"
 
   def zkCandidate = "com.twitter.common.zookeeper" % "candidate" % "0.0.76"