+++ project/LinkerdBuild.scala
@@ -46,6 +46,13 @@ object LinkerdBuild extends Base {
       .withE2e()
       .settings(coverageExcludedPackages := ".*XXX_.*")
 
+    val h2 = projectDir("router/h2")
+      .dependsOn(core)
+      .withTwitterLibs(Deps.finagle("netty4"))
+      .withLibs(Deps.netty4("codec-http2"), Deps.netty4("handler"))
+      .withTests()
+      .withE2e()
+
     val http = projectDir("router/http")
       .dependsOn(core)
       .withTwitterLibs(Deps.finagle("http"))
@@ -72,7 +79,7 @@ object LinkerdBuild extends Base {
 
     val all = projectDir("router")
       .settings(aggregateSettings)
-      .aggregate(core, http, mux, thrift)
+      .aggregate(core, h2, http, mux, thrift)
   }
 
   object Namer {
@@ -160,7 +167,9 @@ object LinkerdBuild extends Base {
        |   -XX:+UseCMSInitiatingOccupancyOnly                \
        |   -XX:CMSInitiatingOccupancyFraction=70             \
        |   -XX:-TieredCompilation                            \
-       |   -XX:+UseStringDeduplication                       "
+       |   -XX:+UseStringDeduplication                       \
+       |   -Dcom.twitter.util.events.sinkEnabled=false       \
+       |   ${LOCAL_JVM_OPTIONS:-}                            "
        |""".stripMargin
 
   object Namerd {
@@ -277,7 +286,7 @@ object LinkerdBuild extends Base {
     )
 
     val BundleProjects = Seq[ProjectReference](
-      Namer.consul, Namer.k8s, Namer.marathon, Namer.serversets,
+      Namer.consul, Namer.k8s, Namer.marathon, Namer.serversets, Namer.zkLeader,
       Storage.etcd, Storage.inMemory, Storage.k8s, Storage.zk, Storage.consul
     )
 
@@ -356,7 +365,7 @@ object LinkerdBuild extends Base {
   object Interpreter {
     val namerd = projectDir("interpreter/namerd")
       .withTests()
-      .dependsOn(Namer.core, Namerd.Iface.interpreterThrift)
+      .dependsOn(Namer.core, Namerd.Iface.interpreterThrift, Router.core)
 
     val fs = projectDir("interpreter/fs")
       .withTests()
@@ -396,9 +405,18 @@ object LinkerdBuild extends Base {
 
     val tls = projectDir("linkerd/tls")
       .dependsOn(core)
+      .withLibs("io.netty" % "netty-tcnative-boringssl-static" % "1.1.33.Fork23")
       .withTests()
 
     object Protocol {
+
+      val h2 = projectDir("linkerd/protocol/h2")
+        .withTests()
+        .withTwitterLibs(Deps.finagle("netty4"))
+        .dependsOn(
+          core,
+          Router.h2)
+
       val http = projectDir("linkerd/protocol/http")
         .withTests().withE2e().withIntegration()
         .withTwitterLibs(Deps.finagle("netty4-http"))
@@ -417,7 +435,7 @@ object LinkerdBuild extends Base {
 
       val all = projectDir("linkerd/protocol")
         .settings(aggregateSettings)
-        .aggregate(http, mux, thrift)
+        .aggregate(h2, http, mux, thrift)
 
       val benchmark = projectDir("linkerd/protocol/benchmark")
         .dependsOn(http, Protocol.http)
@@ -507,7 +525,7 @@ object LinkerdBuild extends Base {
     val BundleProjects = Seq[ProjectReference](
       Namer.consul, Namer.k8s, Namer.marathon, Namer.serversets, Namer.zkLeader, Namer.curator,
       Interpreter.namerd, Interpreter.fs, Interpreter.perHost, Interpreter.k8s,
-      Protocol.mux, Protocol.thrift,
+      Protocol.h2, Protocol.mux, Protocol.thrift,
       Announcer.serversets,
       Telemetry.core, Telemetry.tracelog,
       Tracer.zipkin,
@@ -568,6 +586,7 @@ object LinkerdBuild extends Base {
 
   val router = Router.all
   val routerCore = Router.core
+  val routerH2 = Router.h2
   val routerHttp = Router.http
   val routerMux = Router.mux
   val routerThrift = Router.thrift
@@ -619,6 +638,7 @@ object LinkerdBuild extends Base {
   val linkerdCore = Linkerd.core
   val linkerdMain = Linkerd.main
   val linkerdProtocol = Linkerd.Protocol.all
+  val linkerdProtocolH2 = Linkerd.Protocol.h2
   val linkerdProtocolHttp = Linkerd.Protocol.http
   val linkerdProtocolMux = Linkerd.Protocol.mux
   val linkerdProtocolThrift = Linkerd.Protocol.thrift