+++ router/thrift/src/e2e/scala/io/buoyant/router/ThriftEndToEndTest.scala
@@ -13,8 +13,6 @@ import org.scalatest.FunSuite
 
 class ThriftEndToEndTest extends FunSuite with Awaits {
 
-  override val defaultWait = 5.seconds
-
   case class Downstream(name: String, server: ListeningServer) {
     val address = server.boundAddress.asInstanceOf[InetSocketAddress]
     val port = address.getPort