+++ test-util/src/main/scala/io/buoyant/test/Awaits.scala
@@ -6,7 +6,9 @@ import org.scalatest.exceptions.TestFailedException
 
 trait Awaits {
 
-  def defaultWait: Duration = 2.seconds
+  def defaultWait: Duration =
+    sys.env.get("CI_TERRIBLENESS").map(Duration.parse(_)).getOrElse(2.seconds)
+
   def awaitStackDepth: Int = 4
 
   def await[T](t: Duration)(f: => Future[T]): T =